3.0.56.0 Updated items/buildings and item/building descriptions for Update 56
Cleaned up data and icons a bit, removed unused icons I somehow forgot to after their corresponding items got removed from game

+++ Data/FacilityRecipe.cs
@@ -4,11 +4,11 @@
 	{
 		public List<string> InputNames { get; set; }
 
-		public List<int> InputAmounts { get; set; }
+		public List<float> InputAmounts { get; set; }
 
 		public List<string> OutputNames { get; set; }
 
-		public List<int> OutputAmounts { get; set; }
+		public List<float> OutputAmounts { get; set; }
 
 		public float PowerConsumption { get; set; }
